+++ libs/core/graphql/src/apollo/public_api.ts
@@ -1,2 +1,4 @@
 export { DaffQueuedApollo } from './queued-apollo/queued-apollo.service';
-export * from './cacheable-operations/cacheable-operations';
+export * from './cacheable-operations/cacheable-operations-converter';
+export { DaffApolloCacheableOperationsLinkGenerator } from './cacheable-operations/cacheable-operations-link-generator';
+export { DAFF_APOLLO_CACHEABLE_OPERATIONS } from './cacheable-operations/cacheable-operations-token';
